Refactor selenium tests
